publish everything

+++ src/database/migrations/2019_01_09_173907_create_transactions_table.php
@@ -22,7 +22,7 @@ class CreateTransactionsTable extends Migration
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->string('phone');
-            $table->float('amount');
+            $table->bigInteger('amount');
             $table->text('currency');
 
             //a new transaction will be marked as null until payment is confirmed
